refactor: restructures code structure.
