+++ cmd/plugin/purser_plugin.go
@@ -25,14 +25,34 @@ import (
 
 	"strings"
 
+	"github.com/vmware/purser/pkg/client"
+	groups_client_v1 "github.com/vmware/purser/pkg/client/clientset/typed/groups/v1"
 	"github.com/vmware/purser/pkg/plugin"
-	"github.com/vmware/purser/pkg/plugin/client"
-	"github.com/vmware/purser/pkg/plugin/controller"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
 )
 
-var crdclient *client.Crdclient
+var crdclient *groups_client_v1.GroupClient
+
+func init() {
+	kubeconfig := flag.String("kubeconfig", os.Getenv("KUBECTL_PLUGINS_GLOBAL_FLAG_KUBECONFIG"), "path to Kubernetes config file")
+	flag.Parse()
+
+	config, err := clientcmd.BuildConfigFromFlags("", *kubeconfig)
+	if err != nil {
+		panic(err.Error())
+	}
+
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		panic(err.Error())
+	}
+	plugin.ProvideClientSetInstance(clientset)
+
+	// Crd client
+	client, clusterConfig := client.GetAPIExtensionClient()
+	crdclient = groups_client_v1.NewGroupClient(client, clusterConfig)
+}
 
 func main() {
 	inputs := os.Args[1:]
@@ -40,7 +60,7 @@ func main() {
 	if len(inputs) == 4 && inputs[0] == Get {
 		computeMetricInsight(inputs)
 	} else if len(inputs) == 2 {
-		performAction(inputs)
+		computeStats(inputs)
 	} else {
 		printHelp()
 	}
@@ -55,13 +75,83 @@ func computeMetricInsight(inputs []string) {
 	}
 }
 
-func performAction(inputs []string) {
+func computeCost(inputs []string) {
+	if inputs[2] == Label {
+		plugin.GetPodsCostForLabel(inputs[3])
+	} else if inputs[2] == Pod {
+		plugin.GetPodCost(inputs[3])
+	} else if inputs[2] == Node {
+		plugin.GetAllNodesCost()
+	} else {
+		printHelp()
+	}
+}
+
+func fetchResource(inputs []string) {
+	if inputs[2] == Namespace {
+		group := plugin.GetGroupByName(crdclient, inputs[3])
+		if group != nil {
+			plugin.PrintGroup(group)
+		} else {
+			fmt.Printf("Group %s is not present\n", inputs[3])
+		}
+	} else if inputs[2] == Label {
+		if !strings.Contains(inputs[3], "=") {
+			printHelp()
+		}
+		group := plugin.GetGroupByName(crdclient, createGroupNameFromLabel(inputs[3]))
+		if group != nil {
+			plugin.PrintGroup(group)
+		} else {
+			fmt.Printf("Group %s is not present\n", inputs[3])
+		}
+	} else if inputs[2] == Group {
+		group := plugin.GetGroupByName(crdclient, inputs[3])
+		if group != nil {
+			plugin.PrintGroup(group)
+		} else {
+			fmt.Printf("No group with name: %s\n", inputs[3])
+		}
+	} else {
+		printHelp()
+	}
+}
+
+func createGroupNameFromLabel(input string) string {
+	inp := strings.Split(input, "=")
+	key := inp[0]
+	val := inp[1]
+	groupName := key + "." + val
+	if strings.Contains(groupName, "/") {
+		groupName = strings.Replace(groupName, "/", "-", -1)
+	}
+	groupName = strings.ToLower(groupName)
+	return groupName
+}
+
+func computeStats(inputs []string) {
 	switch inputs[0] {
 	case Get:
 		getStats(inputs)
 	case Set:
 		setStats(inputs)
 	default:
+		printHelp()
+	}
+}
+
+func getStats(inputs []string) {
+	if inputs[1] == "summary" {
+		plugin.GetClusterSummary()
+	} else if inputs[1] == "savings" {
+		plugin.GetSavings()
+	} else if inputs[1] == "user-costs" {
+		price := plugin.GetUserCosts()
+		fmt.Printf("cpu cost per CPU per hour:\t %f$\nmem cost per GB per hour:\t %f$\nstorage cost per GB per hour:\t %f$\n",
+			price.CPU,
+			price.Memory,
+			price.Storage)
+	} else {
 		printHelp()
 	}
 }
@@ -87,95 +177,6 @@ func setStats(inputs []string) {
 	} else {
 		printHelp()
 	}
-}
-
-func getStats(inputs []string) {
-	if inputs[1] == "summary" {
-		plugin.GetClusterSummary()
-	} else if inputs[1] == "savings" {
-		plugin.GetSavings()
-	} else if inputs[1] == "user-costs" {
-		price := plugin.GetUserCosts()
-		fmt.Printf("cpu cost per CPU per hour:\t %f$\nmem cost per GB per hour:\t %f$\nstorage cost per GB per hour:\t %f$\n",
-			price.CPU,
-			price.Memory,
-			price.Storage)
-	} else {
-		printHelp()
-	}
-}
-
-func fetchResource(inputs []string) {
-	if inputs[2] == Namespace {
-		group := controller.GetCrdByName(crdclient, inputs[3])
-		if group != nil {
-			controller.PrintGroup(group)
-		} else {
-			fmt.Printf("Group %s is not present\n", inputs[3])
-		}
-	} else if inputs[2] == Label {
-		if !strings.Contains(inputs[3], "=") {
-			printHelp()
-		}
-		group := controller.GetCrdByName(crdclient, createGroupNameFromLabel(inputs[3]))
-		if group != nil {
-			controller.PrintGroup(group)
-		} else {
-			fmt.Printf("Group %s is not present\n", inputs[3])
-		}
-	} else if inputs[2] == Group {
-		group := controller.GetCrdByName(crdclient, inputs[3])
-		if group != nil {
-			controller.PrintGroup(group)
-		} else {
-			fmt.Printf("No group with name: %s\n", inputs[3])
-		}
-	} else {
-		printHelp()
-	}
-}
-
-func computeCost(inputs []string) {
-	if inputs[2] == Label {
-		plugin.GetPodsCostForLabel(inputs[3])
-	} else if inputs[2] == Pod {
-		plugin.GetPodCost(inputs[3])
-	} else if inputs[2] == Node {
-		plugin.GetAllNodesCost()
-	} else {
-		printHelp()
-	}
-}
-
-func createGroupNameFromLabel(input string) string {
-	inp := strings.Split(input, "=")
-	key := inp[0]
-	val := inp[1]
-	groupName := key + "." + val
-	if strings.Contains(groupName, "/") {
-		groupName = strings.Replace(groupName, "/", "-", -1)
-	}
-	groupName = strings.ToLower(groupName)
-	return groupName
-}
-
-func init() {
-	kubeconfig := flag.String("kubeconfig", os.Getenv("KUBECTL_PLUGINS_GLOBAL_FLAG_KUBECONFIG"), os.Getenv("KUBECTL_PLUGINS_GLOBAL_FLAG_KUBECONFIG"))
-	flag.Parse()
-
-	config, err := clientcmd.BuildConfigFromFlags("", *kubeconfig)
-	if err != nil {
-		panic(err.Error())
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		panic(err.Error())
-	}
-	plugin.ProvideClientSetInstance(clientset)
-
-	// Crd client
-	crdclient = controller.GetAPIExtensionClient()
 }
 
 func printHelp() {
