+++ pkg/apis/groups/v1/deepcopy.go
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2018 VMware Inc. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package v1
+
+import "k8s.io/apimachinery/pkg/runtime"
+
+// DeepCopyInto copies all properties of this object into another object of the
+// same type that is provided as a pointer.
+func (in *Group) DeepCopyInto(out *Group) {
+	out.TypeMeta = in.TypeMeta
+	out.ObjectMeta = in.ObjectMeta
+	out.Spec = in.Spec
+	out.Status = in.Status
+}
+
+// DeepCopyObject returns a generically typed copy of an object
+func (in *Group) DeepCopyObject() runtime.Object {
+	out := Group{}
+	in.DeepCopyInto(&out)
+	return &out
+}
+
+// DeepCopyObject returns a generically typed copy of an object
+func (in *GroupList) DeepCopyObject() runtime.Object {
+	out := GroupList{}
+	out.TypeMeta = in.TypeMeta
+	out.ListMeta = in.ListMeta
+
+	if in.Items != nil {
+		out.Items = make([]*Group, len(in.Items))
+		for i := range in.Items {
+			in.Items[i].DeepCopyInto(out.Items[i])
+		}
+	}
+	return &out
+}
