+++ pkg/apis/groups/v1/docs.go
@@ -1,0 +1,1 @@
+package v1
