--- pkg/controller/controller/payload.go
@@ -1,39 +1,0 @@
-/*
- * Copyright (c) 2018 VMware Inc. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package controller
-
-import (
-	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// PayloadWrapper holds additional information about payload
-type PayloadWrapper struct {
-	OrgID   string         `json:"orgId"`
-	Cluster string         `json:"cluster"`
-	Data    []*interface{} `json:"data"`
-}
-
-// Payload holds payload information
-type Payload struct {
-	Key          string `json:"key"`
-	EventType    string `json:"eventType"`
-	ResourceType string `json:"resourceType"`
-	CloudType    string `json:"cloudType"`
-	Data         string `json:"data"`
-	CaptureTime  meta_v1.Time
-}
