+++ pkg/controller/eventprocessor/processor.go
@@ -1,19 +1,18 @@
 package eventprocessor
 
 import (
+	"encoding/json"
 	"time"
 
-	"encoding/json"
+	"github.com/vmware/purser/pkg/controller"
+	"github.com/vmware/purser/pkg/persistence/dgraph"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/vmware/purser/pkg/controller/config"
-	"github.com/vmware/purser/pkg/controller/controller"
-	"github.com/vmware/purser/pkg/controller/persistence/dgraph"
 	api_v1 "k8s.io/api/core/v1"
 )
 
 // ProcessEvents processes the event and notifies the subscribers.
-func ProcessEvents(conf *config.Config) {
+func ProcessEvents(conf *controller.Config) {
 
 	for {
 		conf.RingBuffer.PrintDetails()
@@ -22,7 +21,7 @@ func ProcessEvents(conf *config.Config) {
 			// TODO: listen for subscriber and group crd updates and update
 			// in memory copy instead of querying everytime.
 			subscribers := getSubscribers(conf)
-			groups := controller.GetAllGroups(conf.Groupcrdclient)
+			groups := getAllGroups(conf.Groupcrdclient)
 
 			data, size := conf.RingBuffer.ReadN(ReadSize)
 
@@ -38,7 +37,7 @@ func ProcessEvents(conf *config.Config) {
 			NotifySubscribers(data, subscribers)
 
 			// Update user created groups.
-			controller.UpdateCustomGroups(data, groups, conf.Groupcrdclient)
+			UpdateCustomGroups(data, groups, conf.Groupcrdclient)
 
 			conf.RingBuffer.RemoveN(size)
 			conf.RingBuffer.PrintDetails()
@@ -47,6 +46,7 @@ func ProcessEvents(conf *config.Config) {
 	}
 }
 
+// PersistPayloads store payload info in dgraph
 func PersistPayloads(payloads []*interface{}) {
 	for _, event := range payloads {
 		payload := (*event).(*controller.Payload)
@@ -58,7 +58,7 @@ func PersistPayloads(payloads []*interface{}) {
 			}
 			err = dgraph.PersistPod(pod)
 			if err != nil {
-				log.Errorf("Error while persisting pod ", err)
+				log.Errorf("Error while persisting pod %v", err)
 			}
 		} else if payload.ResourceType == "Service" {
 			service := api_v1.Service{}
@@ -68,7 +68,7 @@ func PersistPayloads(payloads []*interface{}) {
 			}
 			err = dgraph.PersistService(service)
 			if err != nil {
-				log.Errorf("Error while persisting service ", err)
+				log.Errorf("Error while persisting service %v", err)
 			}
 		}
 	}
