--- pkg/plugin/client/client.go
@@ -1,88 +1,0 @@
-/*
- * Copyright (c) 2018 VMware Inc. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package client
-
-import (
-	"github.com/vmware/purser/pkg/plugin/crd"
-	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/rest"
-)
-
-// This file implement all the (CRUD) client methods we need to access our CRD object
-
-// Crdclient structure
-type Crdclient struct {
-	cl     *rest.RESTClient
-	ns     string
-	plural string
-	codec  runtime.ParameterCodec
-}
-
-// CrdClient creates a new instance of the CRD
-func CrdClient(cl *rest.RESTClient, scheme *runtime.Scheme, namespace string) *Crdclient {
-	return &Crdclient{cl: cl, ns: namespace, plural: crd.CRDPlural,
-		codec: runtime.NewParameterCodec(scheme)}
-}
-
-// Create creates a CRD group.
-func (f *Crdclient) Create(obj *crd.Group) (*crd.Group, error) {
-	var result crd.Group
-	err := f.cl.Post().
-		Namespace(f.ns).Resource(f.plural).
-		Body(obj).Do().Into(&result)
-	return &result, err
-}
-
-// Update updates the CRD group.
-func (f *Crdclient) Update(obj *crd.Group) (*crd.Group, error) {
-	var result crd.Group
-	err := f.cl.
-		//Put().
-		Put().Name(obj.Name).
-		Namespace(f.ns).Resource(f.plural).
-		Body(obj).Do().Into(&result)
-	return &result, err
-}
-
-// Delete deletes the CRD group.
-func (f *Crdclient) Delete(name string, options *meta_v1.DeleteOptions) error {
-	return f.cl.Delete().
-		Namespace(f.ns).Resource(f.plural).
-		Name(name).Body(options).Do().
-		Error()
-}
-
-// Get fetches the CRD group.
-func (f *Crdclient) Get(name string) (*crd.Group, error) {
-	var result crd.Group
-	err := f.cl.Get().
-		Namespace(f.ns).Resource(f.plural).
-		Name(name).Do().Into(&result)
-	return &result, err
-}
-
-// List returns the list of CRD groups.
-func (f *Crdclient) List(opts meta_v1.ListOptions) (*crd.GroupList, error) {
-	var result crd.GroupList
-	err := f.cl.Get().
-		Namespace(f.ns).Resource(f.plural).
-		VersionedParams(&opts, f.codec).
-		Do().Into(&result)
-	return &result, err
-}
