+++ pkg/plugin/costing.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 
 	"github.com/vmware/purser/pkg/plugin/metrics"
-
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
