--- pkg/plugin/crd/group_crd.go
@@ -1,165 +1,0 @@
-/*
- * Copyright (c) 2018 VMware Inc. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package crd
-
-import (
-	"reflect"
-
-	"github.com/vmware/purser/pkg/plugin/metrics"
-
-	apiextv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
-	apiextcs "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
-	"k8s.io/client-go/rest"
-)
-
-const (
-	// CRDPlural plural form value
-	CRDPlural string = "groups"
-	// CRDGroup value
-	CRDGroup string = "vmware.kuber"
-	// CRDVersion value
-	CRDVersion string = "v1"
-	// FullCRDName value
-	FullCRDName string = CRDPlural + "." + CRDGroup
-)
-
-// CreateCRD creates the CRD resource, ignore error if it already exists
-func CreateCRD(clientset apiextcs.Interface) error {
-	crd := &apiextv1beta1.CustomResourceDefinition{
-		ObjectMeta: meta_v1.ObjectMeta{Name: FullCRDName},
-		Spec: apiextv1beta1.CustomResourceDefinitionSpec{
-			Group:   CRDGroup,
-			Version: CRDVersion,
-			//TODO: make cluster scoped?
-			Scope: apiextv1beta1.NamespaceScoped,
-			Names: apiextv1beta1.CustomResourceDefinitionNames{
-				Plural: CRDPlural,
-				Kind:   reflect.TypeOf(Group{}).Name(),
-			},
-		},
-	}
-
-	_, err := clientset.ApiextensionsV1beta1().CustomResourceDefinitions().Create(crd)
-	if err != nil && apierrors.IsAlreadyExists(err) {
-		return nil
-	}
-	return err
-
-	// Note the original apiextensions example adds logic to wait for creation and exception handling
-}
-
-// Group definition of our CRD Group class
-type Group struct {
-	meta_v1.TypeMeta   `json:",inline"`
-	meta_v1.ObjectMeta `json:"metadata"`
-	Spec               GroupSpec   `json:"spec"`
-	Status             GroupStatus `json:"status,omitempty"`
-}
-
-// GroupSpec definition
-type GroupSpec struct {
-	Name               string                      `json:"name"`
-	Type               string                      `json:"type,omitempty"`
-	Labels             map[string]string           `json:"labels,omitempty"`
-	AllocatedResources *metrics.Metrics            `json:"metrics,omitempty"`
-	PodsMetrics        map[string]*metrics.Metrics `json:"pods,omitempty"`
-	PodsDetails        map[string]*PodDetails      `json:"podDetails,omitempty"`
-}
-
-// PodDetails definition
-type PodDetails struct {
-	Name            string
-	StartTime       meta_v1.Time
-	EndTime         meta_v1.Time
-	Containers      []*Container
-	PodVolumeClaims map[string]*PersistentVolumeClaim
-}
-
-// Container definition
-type Container struct {
-	Name    string
-	Metrics *metrics.Metrics
-}
-
-// PersistentVolumeClaim information
-// A PVC can bound and unbound to a pod many times, so maintaining
-// BoundTimes and UnboundTimes as lists.
-// A PVC can be upgraded or downgraded, so maintaining capacityAllocated as a list
-// Whenever a PVC capacity changes will update UnboundTime for old capacity, and
-// append new capacity to capacityAllocated with bound time appended to BoundTimes
-// The i-th capacity alloacted corresponds to the i-th bound time and to i-th unbound time.
-// Similarly for RequestSizeInGB
-type PersistentVolumeClaim struct {
-	Name                string
-	VolumeName          string
-	RequestSizeInGB     []float64
-	CapacityAllotedInGB []float64
-	BoundTimes          []meta_v1.Time
-	UnboundTimes        []meta_v1.Time
-}
-
-// GroupStatus details
-type GroupStatus struct {
-	State   string `json:"state,omitempty"`
-	Message string `json:"message,omitempty"`
-}
-
-// GroupList information
-type GroupList struct {
-	meta_v1.TypeMeta `json:",inline"`
-	meta_v1.ListMeta `json:"metadata"`
-	Items            []Group `json:"items"`
-}
-
-// SchemeGroupVersion creates a Rest client with the new CRD Schema
-var SchemeGroupVersion = schema.GroupVersion{Group: CRDGroup, Version: CRDVersion}
-
-func addKnownTypes(scheme *runtime.Scheme) error {
-	scheme.AddKnownTypes(SchemeGroupVersion,
-		&Group{},
-		&GroupList{},
-	)
-	meta_v1.AddToGroupVersion(scheme, SchemeGroupVersion)
-	return nil
-}
-
-// NewClient returns a new instance of REST client.
-func NewClient(cfg *rest.Config) (*rest.RESTClient, *runtime.Scheme, error) {
-	scheme := runtime.NewScheme()
-	SchemeBuilder := runtime.NewSchemeBuilder(addKnownTypes)
-	if err := SchemeBuilder.AddToScheme(scheme); err != nil {
-		return nil, nil, err
-	}
-	config := *cfg
-	config.GroupVersion = &SchemeGroupVersion
-	config.APIPath = "/apis"
-	config.ContentType = runtime.ContentTypeJSON
-	config.NegotiatedSerializer = serializer.DirectCodecFactory{
-		CodecFactory: serializer.NewCodecFactory(scheme)}
-
-	client, err := rest.RESTClientFor(&config)
-	if err != nil {
-		return nil, nil, err
-	}
-	return client, scheme, nil
-}
